add the option to package the AOT-compiled classes into a jar
instead of bundling them as classes in WEB-INF/classes.
This is important for those with fairly large projects (or projects that pull in a lot of dependencies that get AOT-compiled).
This can actually lead to too many open files depending on the file descriptor limits of your OS.  Yes, you can bump the FD
limits at the OS level, but this is much friendlier.
This does invoke the 'lein jar' task behind the scenes, fyi.