print any exceptions from require-and-resolve
tomcat doesn't log tracebacks if there's a failure in the listener init
method, it just stops. Once we're in the user's code, the user can log
errors themself. But if there's an error that arises while we're
compiling and loading the user's code, the user has no way to catch and
log it, so we should probably give them some chance of knowing what
happened